+++ expire/newsrequeue.c
@@ -344,7 +344,7 @@ ParseNewsfeeds(name)
 		break;
 	    }
 	    if (*p == '\\' && p[1] == '\n')
-		while (*++p && CTYPE(isspace, *p))
+		while (*++p && CTYPE(isspace, (int)*p))
 		    continue;
 	    else
 		*to++ = *p++;
